# openapi.yaml
openapi: 3.0.3
info:
  title: Billr API
  description: |
    B2B subscription billing API. Manage accounts, plans, subscriptions, usage metering,
    invoices, and webhooks. All monetary values are integers in the smallest currency unit
    (e.g. cents for USD). Never floats.

    ## Authentication
    All endpoints require a Bearer API key in the `Authorization` header, except
    `GET /plans` and `GET /plans/{planId}` which are public.

    ## Rate Limiting
    Requests are rate-limited per API key. Limits are returned in response headers:
    - `X-RateLimit-Limit` — requests allowed per window
    - `X-RateLimit-Remaining` — requests remaining
    - `X-RateLimit-Reset` — UTC epoch seconds when window resets

    ## Pagination
    List endpoints use cursor-based pagination. Pass `after` (a record ID) and `limit`
    (default 20, max 100). If `next_cursor` is present in the response there are more pages.

    ## Idempotency
    POST endpoints that create resources support an `Idempotency-Key` request header.
    Supply a unique UUID per logical operation; duplicate keys within 24 h return the
    cached response without re-executing the operation.
  version: 1.0.0
  contact:
    name: Billr Support
    email: support@billr.io
    url: https://docs.billr.io
  license:
    name: Proprietary
servers:
  - url: https://api.billr.io/v1
    description: Production
security:
  - BearerAuth: []
...
paths:
...
  /auth/keys:
    post:
      tags: [Auth]
      operationId: createApiKey
      summary: Create API key
      description: |
        Creates a new API key. The `key` field is only returned once in this response;
        it cannot be retrieved afterwards. Store it securely immediately.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApiKeyRequest'
            example:
              description: "Production server key"
              scope: admin
      responses:
        '201':
          description: API key created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateApiKeyResponse'
              example:
                id: key_01HX3Y9ZG2VMBPN5Q8T7RWKD
                key: bllr_live_sk_MzQ2YjFkNzItNDc4Zi00OTY3LWI4YTUtNjY2ZWM5NjA1Zjdl
                prefix: bllr_live_sk_MzQ2
                scope: admin
                created_at: "2026-04-19T10:00:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'

    get:
      tags: [Auth]
      operationId: listApiKeys
      summary: List API keys
      description: Returns all API keys for the authenticated workspace. Key values are never returned here — only metadata.
      responses:
        '200':
          description: List of API keys
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ApiKey'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/keys/{keyId}:
    delete:
      tags: [Auth]
      operationId: deleteApiKey
      summary: Revoke API key
      description: Permanently revokes the API key. In-flight requests using this key will fail immediately.
      parameters:
        - $ref: '#/components/parameters/KeyId'
      responses:
        '204':
          description: Key revoked
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ---------------------------------------------------------------------------
  # Accounts
  # ---------------------------------------------------------------------------
...
  /plans:
    get:
      tags: [Plans]
      operationId: listPlans
      summary: List plans
      description: Returns all active billing plans. This endpoint is public — no authentication required.
      security: []
      responses:
        '200':
          description: Plan list
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Plan'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags: [Plans]
      operationId: createPlan
      summary: Create plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePlanRequest'
            example:
              name: "Growth"
              slug: "growth"
              billing_interval: monthly
              base_price: 34900
              trial_days: 14
              stripe_price_id: "price_1NkXXXXXXX"
              description: "For scaling SaaS teams"
      responses:
        '201':
          description: Plan created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plan'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /plans/{planId}:
    get:
      tags: [Plans]
      operationId: getPlan
      summary: Get plan
      description: Returns a single plan. Public — no authentication required.
      security: []
      parameters:
        - $ref: '#/components/parameters/PlanId'
      responses:
        '200':
          description: Plan object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plan'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    patch:
      tags: [Plans]
      operationId: updatePlan
      summary: Update plan
      description: |
        Partial update. Changing `base_price` does not affect existing subscriptions —
        they continue at the price locked at subscription creation. Only new subscriptions
        pick up the new price.
      parameters:
        - $ref: '#/components/parameters/PlanId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePlanRequest'
      responses:
        '200':
          description: Updated plan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plan'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags: [Plans]
      operationId: deletePlan
      summary: Archive plan
      description: Archives the plan so it is no longer offered to new subscribers. Existing subscriptions are not affected.
      parameters:
        - $ref: '#/components/parameters/PlanId'
      responses:
        '204':
          description: Plan archived
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ---------------------------------------------------------------------------
  # Subscriptions
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: API key obtained from POST /auth/keys. Prefix with `Bearer `.
  parameters:
...
    PlanId:
      name: planId
      in: path
      required: true
      schema:
        type: string
      description: Plan ID (prefix `plan_`)
...
    KeyId:
      name: keyId
      in: path
      required: true
      schema:
        type: string
      description: API key ID (prefix `key_`)
...
  responses:
    BadRequest:
      description: Invalid request body or query parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: validation_error
            message: "'billing_interval' must be one of: monthly, yearly"
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: unauthorized
            message: "Invalid or missing Bearer token"
    Forbidden:
      description: API key does not have permission for this operation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: forbidden
            message: "read_only keys cannot perform this action"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: not_found
            message: "Account acc_01HX3Y9ZG2VMBPN5Q8T7RWKD not found"
    TooManyRequests:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds until the rate limit resets
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: rate_limited
            message: "Rate limit exceeded. Retry after 30 seconds."
    InternalServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: internal_error
            message: "An unexpected error occurred. Our team has been notified."
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code
          example: not_found
        message:
          type: string
          description: Human-readable description
          example: "Account not found"

    # -------------------------------------------------------------------------
    # Auth
    # -------------------------------------------------------------------------
    ApiKey:
      type: object
      required: [id, prefix, scope, created_at]
      properties:
        id:
          type: string
          example: key_01HX3Y9ZG2VMBPN5Q8T7RWKD
        prefix:
          type: string
          description: First 16 characters of the key — safe to display
          example: bllr_live_sk_MzQ2
        scope:
          $ref: '#/components/schemas/ApiKeyScope'
        description:
          type: string
          nullable: true
          example: "Production server key"
        created_at:
          type: string
          format: date-time
        last_used_at:
          type: string
          format: date-time
          nullable: true
    ApiKeyScope:
      type: string
      enum: [admin, read_only]
      description: "`admin` has full access. `read_only` can only perform GET requests."
    CreateApiKeyRequest:
      type: object
      required: [scope]
      properties:
        description:
          type: string
          maxLength: 255
        scope:
          $ref: '#/components/schemas/ApiKeyScope'
    CreateApiKeyResponse:
      type: object
      required: [id, key, prefix, scope, created_at]
      properties:
        id:
          type: string
        key:
          type: string
          description: Full API key value — shown exactly once. Store immediately.
        prefix:
          type: string
        scope:
          $ref: '#/components/schemas/ApiKeyScope'
        created_at:
          type: string
          format: date-time

    # -------------------------------------------------------------------------
    # Accounts
    # -------------------------------------------------------------------------
...
    Plan:
      type: object
      required: [id, name, slug, billing_interval, base_price, is_active, created_at, updated_at]
      properties:
        id:
          type: string
          example: plan_01HX3Y9ZG2VMBPN5Q8T7RWKD
        name:
          type: string
          example: Growth
        slug:
          type: string
          description: URL-safe identifier
          example: growth
        description:
          type: string
          nullable: true
        billing_interval:
          $ref: '#/components/schemas/BillingInterval'
        base_price:
          type: integer
          description: Price in smallest currency unit (e.g. cents). Never a float.
          example: 34900
        trial_days:
          type: integer
          minimum: 0
          default: 0
          example: 14
        stripe_price_id:
          type: string
          nullable: true
          example: price_1NkXXXXXXX
        tiers:
          type: array
          items:
            $ref: '#/components/schemas/PlanTier'
          description: Optional volume/tiered pricing tiers. Empty for flat-rate plans.
        is_active:
          type: boolean
          description: Archived plans are not shown publicly
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    PlanTier:
      type: object
      required: [up_to, unit_amount]
      description: A single pricing tier for volume or graduated billing.
      properties:
        up_to:
          type: integer
          nullable: true
          description: Inclusive upper bound for this tier. `null` means unlimited (last tier).
          example: 10000
        unit_amount:
          type: integer
          description: Price per unit in smallest currency unit within this tier
          example: 5
        flat_amount:
          type: integer
          nullable: true
          description: Flat fee charged when this tier is entered (graduated model only)
          example: 0
    BillingInterval:
      type: string
      enum: [monthly, yearly]
    CreatePlanRequest:
      type: object
      required: [name, slug, billing_interval, base_price]
      properties:
        name:
          type: string
        slug:
          type: string
          pattern: '^[a-z0-9-]+$'
        billing_interval:
          $ref: '#/components/schemas/BillingInterval'
        base_price:
          type: integer
          minimum: 0
        trial_days:
          type: integer
          minimum: 0
          default: 0
        stripe_price_id:
          type: string
        description:
          type: string
    UpdatePlanRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        trial_days:
          type: integer
          minimum: 0
        stripe_price_id:
          type: string
        is_active:
          type: boolean

    # -------------------------------------------------------------------------
    # Subscriptions
    # -------------------------------------------------------------------------
...